+++ src/equicordplugins/globalBadges/index.tsx
@@ -13,9 +13,8 @@ import { BadgeContextMenu } from "@plugins/_api/badges";
 import { Devs, EquicordDevs } from "@utils/constants";
 import { openInviteModal } from "@utils/discord";
 import definePlugin from "@utils/types";
-import { ContextMenuApi, React, Toasts, UserStore } from "@webpack/common";
+import { ContextMenuApi, React, Toasts } from "@webpack/common";
 
-import { openBadgeModal } from "./badgeModal";
 import { settings } from "./settings";
 import { cl, GlobalBadges, INVITE_LINK, loadBadges } from "./utils";
 
@@ -74,9 +73,6 @@ export default definePlugin({
             },
             onContextMenu(event, badge) {
                 ContextMenuApi.openContextMenu(event, () => <BadgeContextMenu badge={badge} />);
-            },
-            onClick() {
-                return openBadgeModal(UserStore.getUser(userId));
             },
         } satisfies ProfileBadge));
     }
