+++ src/equicordplugins/globalBadges/styles.css
@@ -1,49 +1,3 @@
-.vc-global-badges-modal-name {
-    text-transform: none;
-    flex-grow: 0;
-    background: var(--background-base-lowest);
-    border-radius: 0 9999px 9999px 0;
-    padding: 6px 0.8em 6px 0.5em;
-    font-size: 20px;
-    height: 20px;
-    position: relative;
-    text-wrap: nowrap;
-}
-
-.vc-global-badges-modal-name::before {
-    content: "";
-    display: block;
-    position: absolute;
-    height: 100%;
-    width: 32px;
-    background: var(--background-base-lowest);
-    z-index: -1;
-    left: -32px;
-    top: 0;
-    border-top-left-radius: 9999px;
-    border-bottom-left-radius: 9999px;
-}
-
-.vc-global-badges-modal-avatar {
-    height: 48px;
-    width: 48px;
-    border-radius: 50%;
-}
-
-.vc-global-badges-modal-description {
-    font-size: 16px;
-}
-
-.vc-global-badges-modal-text {
-    margin-left: 8px;
-}
-
-.vc-global-badges-modal-badge {
-    width: 50px;
-    height: 50px;
-    margin: 2px;
-}
-
 .vc-global-badges-settings-button {
     width: 100%;
     height: 50px;
